requirements.txt
sentence-transformers
umap-learn
torchinfo